ewallet_api/requirements.txt
asgiref==3.5.2
autopep8==1.6.0
certifi==2022.6.15
click==8.1.3
colorama==0.4.5
distlib==0.3.5
Django==4.0.6
djangorestframework==3.13.1
filelock==3.7.1
Flask==2.1.3
Flask-Cors==3.0.10
Flask-MySQL==1.5.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysqlclient==2.1.1
pipenv==2022.7.24
platformdirs==2.5.2
pycodestyle==2.9.0
PyMySQL==1.0.2
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tzdata==2022.1
virtualenv==20.16.2
virtualenv-clone==0.5.7
Werkzeug==2.2.1
